--- index.js
@@ -1,4 +1,0 @@
-export * from "./src/formatBytes.js";
-export * from "./src/locationToString.js";
-export * from "./src/taggedString.js";
-export * from "./src/parseManifestAuthor.js";
