+++ package.json
@@ -11,8 +11,10 @@
     "test": "test"
   },
   "scripts": {
-    "lint": "eslint index.js",
-    "test-only": "node --test test/",
+    "build": "tsc",
+    "prepublishOnly": "npm run build",
+    "lint": "eslint src/**/*.ts test/**/*.ts",
+    "test-only": "glob -c \"tsx --test\" \"./test/**/*.spec.ts\"",
     "test": "npm run lint && npm run test-only",
     "coverage": "c8 -r html npm test"
   },
@@ -42,6 +44,8 @@
     "eslint": "^8.34.0"
   },
   "dependencies": {
-    "@npm/types": "^1.0.2"
+    "glob": "^10.3.10",
+    "tsx": "^4.6.2",
+    "typescript": "^4.9.5"
   }
 }
