--- src/taggedString.js
@@ -1,14 +1,0 @@
-export function taggedString(chaines, ...cles) {
-  return function cur(...valeurs) {
-    const dict = valeurs[valeurs.length - 1] || {};
-    const resultat = [chaines[0]];
-    cles.forEach((cle, index) => {
-      resultat.push(
-        typeof cle === "number" ? valeurs[cle] : dict[cle],
-        chaines[index + 1]
-      );
-    });
-
-    return resultat.join("");
-  };
-}
