+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export interface Person {
+    name: string;
+    email?: string;
+    url?: string;
+  }
