+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+    "compilerOptions": {
+      "declaration": true,
+      "strictNullChecks": true,
+      "noImplicitThis": true,
+      "noEmitOnError": true,
+      "noImplicitAny": true,
+      "target": "ES2022",
+      "outDir": "dist",
+      "module": "ES2022",
+      "moduleResolution": "node",
+    //   "esModuleInterop": true,
+    //   "resolveJsonModule": true,
+      "skipDefaultLibCheck": true,
+      "forceConsistentCasingInFileNames": true,
+    //   "sourceMap": true,
+      "rootDir": "./src",
+      // "types": ["node"],
+    },
+    "include": ["src"],
+    "exclude": ["node_modules", "dist"]
+}
